requirements.txt
echo "fastapi==0.110.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
anthropic==0.18.0
openai==1.12.0
pydub==0.25.1
ffmpeg-python==0.2.0
requests==2.31.0" > requirements.txt

pip install -r requirements.txt